Code Reorganization
Next files were created to split the whole schema logic in modules: complaint_types, auth_types, complaint_queries, auth_queries, complaing_resolvers and auth_resolvers.

closes #24
